requirements.txt

python-telegram-bot==20.3
pillow
opencv-python-headless
pytesseract
matplotlib
